LTR helper in engine.solr.ltr
