+++ docker/data-science/notebooks/ltr/log.py
@@ -61,7 +61,7 @@ class FeatureLogger:
 
             ids = [str(doc_id) for doc_id in ids]
             #collection = self.engine.get_collection(self.index)
-            res = self.engine.log_query(self.index, self.feature_set, ids,
+            res = self.ltr.log_query(self.index, self.feature_set, ids,
                                         params, self.id_field)
 
             # Add feature back to each judgment
